Can trigger specific card effects now + card can enter board
Make the SpecificCardEffect inherit from the CardEffect class and overload its cardEffect function so now runEffects refers to CardEffect.cardEffect() but on each card cardEffect() is different.

The cards' parent is board1 when you drag and click it on the board even though it doesn't say it in the hierarchy.