Scrapped everythingBefore
Apparently you need to Instantiate everything individually in order to make a deep copy instead of just instantiating the parent. Doing that crashes Unity though so I decided to just instantiate Before as a copy of After.The Before-cards get their effects triggered even when (or just before) being deactiviated. So I added an exception to that. Should probably make them unsubscribe to the runEffects.